added filters for the todo lists
